requirements.txt
langchain_core
python-dotenv
streamlit
langchain_community
chromadb
groq
langchain-groq
sentence_transformers
langchain-objectbox
langchain
langchain-huggingface
validators
langchain_chroma
youtube-transcript-api
pytube